requirements.txt
numpy
opencv-python
Pillow
scikit-learn
tensorflow
keras